requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
requests==2.32.3
redis==5.0.8
prometheus-fastapi-instrumentator==7.0.0
scikit-learn==1.5.2
joblib==1.4.2
pydantic-settings==2.6.1
pytest==8.3.3
httpx==0.27.2
